Updated to support AES as a hasher.
The idea being if you are working with a remote api that needs to validate your ldap credentials. In this situation there is no way to do a Hash::check($plain, $hashed) on the remote api side, as you do not want to send your password in plain text.

You need to update your config for app.php, and hash.php. And also register the new provider.